chore: start.sh 수정

+++ start.sh
@@ -17,4 +17,8 @@ do
     echo "새로운 어플리케이션을 실행 중입니다.";
 done
 
-kill -15 $(lsof -ti:$blue)
+sleep 5m
+
+if [ -n "$(lsof -ti:$blue)" ]; then
+  kill -15 $(lsof -ti:$blue)
+fi
